Move CoordinateMapping to avoid cyclic imports
