+++ bokeh/models/canvas.py
@@ -38,45 +38,11 @@ from .scales import LinearScale, Scale
 __all__ = (
     "Canvas",
     "CartesianFrame",
-    "CoordinateMapping",
 )
 
 #-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
-
-class CoordinateMapping(Model):
-    """ A mapping between two coordinate systems. """
-
-    # explicit __init__ to support Init signatures
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-    x_source = Instance(Range, default=InstanceDefault(DataRange1d), help="""
-    The source range of the horizontal dimension of the new coordinate space.
-    """)
-
-    y_source = Instance(Range, default=InstanceDefault(DataRange1d), help="""
-    The source range of the vertical dimension of the new coordinate space.
-    """)
-
-    x_scale = Instance(Scale, default=InstanceDefault(LinearScale), help="""
-    What kind of scale to use to convert x-coordinates from the source (data)
-    space into x-coordinates in the target (possibly screen) coordinate space.
-    """)
-
-    y_scale = Instance(Scale, default=InstanceDefault(LinearScale), help="""
-    What kind of scale to use to convert y-coordinates from the source (data)
-    space into y-coordinates in the target (possibly screen) coordinate space.
-    """)
-
-    x_target = Instance(Range, help="""
-    The horizontal range to map x-coordinates in the target coordinate space.
-    """)
-
-    y_target = Instance(Range, help="""
-    The vertical range to map y-coordinates in the target coordinate space.
-    """)
 
 class CartesianFrame(Model):
     """ """
