+++ bokeh/models/coordinates.py
@@ -4,18 +4,6 @@
 #
 # The full license is in the file LICENSE.txt, distributed with this software.
 #-----------------------------------------------------------------------------
-''' Provide Bokeh model "building block" classes.
-
-One of the central design principals of Bokeh is that, regardless of
-how the plot creation code is spelled in Python (or other languages),
-the result is an object graph that encompasses all the visual and
-data aspects of the scene. Furthermore, this *scene graph* is to be
-serialized, and it is this serialized graph that the client library
-BokehJS uses to render the plot. The low-level objects that comprise
-a Bokeh scene graph are called :ref:`Models <bokeh.model>`.
-
-'''
-# This file is excluded from flake8 checking in setup.cfg
 
 #-----------------------------------------------------------------------------
 # Boilerplate
@@ -30,48 +18,55 @@ log = logging.getLogger(__name__)
 #-----------------------------------------------------------------------------
 
 # Bokeh imports
+from ..core.properties import Instance, InstanceDefault
 from ..model import Model
-from .annotations import *  # lgtm [py/polluting-import]
-from .axes import *
-from .callbacks import *
-from .canvas import *
-from .coordinates import *
-from .contour_renderer import *
-from .expressions import *
-from .filters import *
-from .formatters import *
-from .glyphs import *
-from .graphs import *
-from .grids import *
-from .labeling import *
-from .layouts import *
-from .map_plots import *
-from .mappers import *
-from .plots import *
-from .ranges import *
-from .renderers import *
-from .scales import *
-from .selections import *
-from .selectors import *
-from .sources import *
-from .text import *
-from .textures import *
-from .tickers import *
-from .tiles import *
-from .tools import *
-from .transforms import *
-from .ui import *
-from .widgets import *  # lgtm [py/polluting-import]
+from .ranges import DataRange1d, Range
+from .scales import LinearScale, Scale
 
 #-----------------------------------------------------------------------------
 # Globals and constants
 #-----------------------------------------------------------------------------
 
-# __all__ = include all explicit transitive imports above
+__all__ = (
+    "CoordinateMapping",
+)
 
 #-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
+
+class CoordinateMapping(Model):
+    """ A mapping between two coordinate systems. """
+
+    # explicit __init__ to support Init signatures
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+
+    x_source = Instance(Range, default=InstanceDefault(DataRange1d), help="""
+    The source range of the horizontal dimension of the new coordinate space.
+    """)
+
+    y_source = Instance(Range, default=InstanceDefault(DataRange1d), help="""
+    The source range of the vertical dimension of the new coordinate space.
+    """)
+
+    x_scale = Instance(Scale, default=InstanceDefault(LinearScale), help="""
+    What kind of scale to use to convert x-coordinates from the source (data)
+    space into x-coordinates in the target (possibly screen) coordinate space.
+    """)
+
+    y_scale = Instance(Scale, default=InstanceDefault(LinearScale), help="""
+    What kind of scale to use to convert y-coordinates from the source (data)
+    space into y-coordinates in the target (possibly screen) coordinate space.
+    """)
+
+    x_target = Instance(Range, help="""
+    The horizontal range to map x-coordinates in the target coordinate space.
+    """)
+
+    y_target = Instance(Range, help="""
+    The vertical range to map y-coordinates in the target coordinate space.
+    """)
 
 #-----------------------------------------------------------------------------
 # Dev API
